+++ backend/src/baserow/contrib/database/api/formula/views.py
@@ -80,6 +80,6 @@ class TypeFormulaView(APIView):
         field.formula = data["formula"]
         typed_table = type_table(field.table, overridden_field=field)
         # noinspection PyTypeChecker
-        typed_field: FormulaField = typed_table.get_typed_field_instance(field.id)
+        typed_field: FormulaField = typed_table.get_typed_field_instance(field.name)
 
         return Response(TypeFormulaResultSerializer(typed_field).data)