+++ backend/src/baserow/contrib/database/application_types.py
@@ -2,14 +2,11 @@ from django.core.management.color import no_style
 from django.db import connection
 from django.urls import path, include
 
+from baserow.contrib.database.api.serializers import DatabaseSerializer
 from baserow.contrib.database.fields.registries import field_type_registry
+from baserow.contrib.database.models import Database, Table
 from baserow.contrib.database.views.registries import view_type_registry
 from baserow.core.registries import ApplicationType
-from baserow.contrib.database.api.serializers import DatabaseSerializer
-from baserow.contrib.database.formula.types.typed_field_updater import (
-    type_table_and_update_fields,
-)
-from baserow.contrib.database.models import Database, Table
 from baserow.core.trash.handler import TrashHandler
 
 
@@ -145,17 +142,12 @@ class DatabaseApplicationType(ApplicationType):
                     table["_object"], view, id_mapping, files_zip, storage
                 )
 
-            # Once all the fields have been deserialized and created we have to ensure
-            # all fields have been typed and their formulas have correctly been changed
-            # from containing field('..') to field_by_id(..).
-            typed_table = type_table_and_update_fields(table["_object"])
             # We don't need to create all the fields individually because the schema
             # editor can handle the creation of the table schema in one go.
             with connection.schema_editor() as schema_editor:
                 model = table["_object"].get_model(
                     fields=table["_field_objects"],
                     field_ids=[],
-                    typed_table=typed_table,
                 )
                 table["_model"] = model
                 schema_editor.create_model(model)