+++ changelog.md
@@ -11,6 +11,7 @@
 * Added a licensing system for the premium version.
 * Fixed bug where it was possible to create duplicate trash entries. 
 * Fixed propType validation error when converting from a date field to a boolean field.
+* Deprecate internal formula field function field_by_id.
 
 ## Released (2021-10-05)